Make sure cachdir is reset all the time
