+++ tests/conftest.py
@@ -1,3 +1,4 @@
+from unittest.mock import patch
 import pytest
 
 
@@ -7,3 +8,13 @@ def setup_before_test():
     from abstract_ranker.local_llms import reset
 
     reset()
+
+
+@pytest.fixture(autouse=True)
+def cache_dir(tmp_path):
+    # Create a temporary test directory
+    cache_dir = tmp_path / "cache_dir"
+    cache_dir.mkdir()
+
+    with patch("abstract_ranker.config.CACHE_DIR", cache_dir):
+        yield cache_dir
