+++ tests/test_indico.py
@@ -2,18 +2,6 @@ import json
 from pathlib import Path
 from unittest.mock import patch
 
-import pytest
-
-
-@pytest.fixture
-def cache_dir(tmp_path):
-    # Create a temporary test directory
-    cache_dir = tmp_path / "cache_dir"
-    cache_dir.mkdir()
-
-    with patch("abstract_ranker.config.CACHE_DIR", cache_dir):
-        yield cache_dir
-
 
 def test_good_load(cache_dir):
     expected_url = (
